+++ Health/src/main/res/layout/app_bar_main.xml
@@ -21,17 +21,14 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_main" />
-
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
         android:layout_margin="@dimen/fab_margin"
-        android:src="@drawable/ic_add_white" />
-        android:src="@drawable/ic_plus"
-        android:clickable="true"/>
+        android:src="@drawable/ic_add_white"
+        android:clickable="true" />
 
 
 </android.support.design.widget.CoordinatorLayout>
